+++ test/res.locals.js
@@ -3,15 +3,12 @@ var express = require('../')
   , request = require('./support/http');
 
 describe('res', function(){
-  describe('.locals(obj)', function(){
-    it('should merge locals', function(done){
+  describe('.locals', function(){
+    it('should be empty by default', function(done){
       var app = express();
 
       app.use(function(req, res){
         Object.keys(res.locals).should.eql([]);
-        res.locals({ user: 'tobi', age: 1 });
-        res.locals.user.should.equal('tobi');
-        res.locals.age.should.equal(1);
         res.end();
       });