+++ History.md
@@ -10,6 +10,7 @@
    - `req.accepts*` -> `req.accepts*s` - i.e. `req.acceptsEncoding` -> `req.acceptsEncodings`
    - `req.params` is now an object instead of an array
    - `json spaces` no longer enabled by default in development
+   - `res.locals` is no longer a function. It is a plain js object. Treat it as such.
  * refactor:
    - `req.accepts*` with [accepts](https://github.com/expressjs/accepts)
    - `req.is` with [type-is](https://github.com/expressjs/type-is)